+++ package.json
@@ -15,6 +15,7 @@
     "@chakra-ui/react": "^2.8.2",
     "@emotion/react": "^11.13.0",
     "@emotion/styled": "^11.13.0",
+    "@fortawesome/fontawesome-free": "^6.7.2",
     "@uidotdev/usehooks": "^2.4.1",
     "firebase": "^10.12.5",
     "framer-motion": "^11.3.24",
