+++ src/app/layout.tsx
@@ -9,6 +9,9 @@ const inter = Inter({ subsets: ['latin'] });
 export const metadata: Metadata = {
     title: 'Portfolio Web Page',
     description: "Metin Nabiyev's Portfolio Web Page",
+    icons: {
+    icon: '/logo.png',
+  },
 };
 
 export default function RootLayout({
@@ -19,10 +22,6 @@ export default function RootLayout({
     return (
         <ViewTransitions>
             <html lang="en">
-                <link
-                    rel="stylesheet"
-                    href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css"
-                />
                 <body className={inter.className}>
                     <div className="App">
                         <div className="container-app">
